numvariants vs AAF for mix1,... and snp|indel

+++ src/truth_sets_aaf.py
@@ -1,9 +1,10 @@
+import pandas as pd
 import shutil
 import subprocess
 import os
 import joint_gt_ceph as jgc
 
-def make_ts_aaf(mix='mix1',
+def make_ts_aaf(mix='mix1', vartype='snp',
         tsdir='/home/attila/projects/bsm/results/2019-03-18-truth-sets/chr22/snp/truthset'):
         def helper(aaf):
             genotypes = gt_of_aaf[mix][aaf]
@@ -13,10 +14,11 @@ def make_ts_aaf(mix='mix1',
             args0 = ['bcftools', 'concat', '-o', unsorted_outvcf, '-Oz'] + invcfs
             args1 = ['bcftools', 'sort', '-o', outvcf, '-Oz', unsorted_outvcf]
             args2 = ['bcftools', 'index', '-t', outvcf]
-            subprocess.run(args0)
-            subprocess.run(args1)
-            subprocess.run(args2)
-            os.remove(unsorted_outvcf)
+            if not os.path.isfile(outvcf):
+                subprocess.run(args0)
+                subprocess.run(args1)
+                os.remove(unsorted_outvcf)
+                subprocess.run(args2)
             # count records
             args3 = ['bcftools', 'view', '-H', outvcf]
             args4 = ['wc', '-l']
@@ -39,5 +41,7 @@ def make_ts_aaf(mix='mix1',
             os.mkdir(mixdir)
         aafs = gt_of_aaf[mix].keys()
         nrec = [helper(f) for f in aafs]
-        res = {'aaf': aafs, 'nvariants': nrec}
+        res = pd.DataFrame({'vartype': vartype, 'sample': mix, 'aaf': list(aafs),
+                'nvariants': nrec})
+        res = res.astype({'vartype': 'category', 'sample': 'category'})
         return(res)
